+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/blocks/EmptyBlockCheckTest.java
@@ -226,14 +226,22 @@ public class EmptyBlockCheckTest
     }
 
     @Test
-    public void testEmptyBlockSwitchExpressions() throws Exception {
+    public void testEmptyBlockSwitchExpressionsOne() throws Exception {
         final String[] expected = {
             "17:30: " + getCheckMessage(MSG_KEY_BLOCK_NO_STATEMENT, "default"),
-            "116:32: " + getCheckMessage(MSG_KEY_BLOCK_NO_STATEMENT, "case"),
-            "118:26: " + getCheckMessage(MSG_KEY_BLOCK_NO_STATEMENT, "case"),
         };
         verifyWithInlineConfigParser(
-                getNonCompilablePath("InputEmptyBlockSwitchExpressions.java"), expected);
+                getNonCompilablePath("InputEmptyBlockSwitchExpressionsOne.java"), expected);
+    }
+
+    @Test
+    public void testEmptyBlockSwitchExpressionsTwo() throws Exception {
+        final String[] expected = {
+            "25:32: " + getCheckMessage(MSG_KEY_BLOCK_NO_STATEMENT, "case"),
+            "27:26: " + getCheckMessage(MSG_KEY_BLOCK_NO_STATEMENT, "case"),
+        };
+        verifyWithInlineConfigParser(
+                getNonCompilablePath("InputEmptyBlockSwitchExpressionsTwo.java"), expected);
     }
 
     @Test
