+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/blocks/emptyblock/InputEmptyBlockSwitchExpressionsOne.java
@@ -9,7 +9,7 @@ tokens = LITERAL_DEFAULT, LITERAL_CASE, LITERAL_SWITCH
 //non-compiled with javac: Compilable with Java17
 package com.puppycrawl.tools.checkstyle.checks.blocks.emptyblock;
 
-public class InputEmptyBlockSwitchExpressions {
+public class InputEmptyBlockSwitchExpressionsOne {
     void howMany1(Nums k) {
         switch (k) {
             case ONE:
@@ -98,26 +98,6 @@ public class InputEmptyBlockSwitchExpressions {
                 break;
             default:
                 throw new IllegalStateException("Not a nums");
-        }
-    }
-
-    void howMany7(Nums k) {
-        switch (k) {
-            case ONE -> System.out.println("case one");
-            case TWO, THREE -> { System.out.println("case two");}
-            case FOUR -> System.out.println("case three");
-            default -> throw new IllegalStateException("Not a nums");
-        }
-    }
-
-    void howMany8(Nums k) {
-        switch (k) {
-            case ONE -> System.out.println("case two");
-            case TWO, THREE -> {} // violation, 'Must have at least one statement'
-
-            case FOUR -> {} // violation, 'Must have at least one statement'
-
-            default -> throw new IllegalStateException("Not a nums");
         }
     }
 }
